requirements.txt
pocketflow>=0.0.1
python-dotenv>=1.0.0
langchain-openai>=0.1.0
anthropic>=0.25.0
ollama>=0.1.0
google-generativeai
pyyaml>=6.0